Make setting global since filter logic does not support on per server basis
