+++ src/main/java/com/sonyericsson/hudson/plugins/gerrit/trigger/GerritServer.java
@@ -565,8 +565,6 @@ public class GerritServer implements Describable<GerritServer>, Action {
                 missedEventsPlaybackManager.checkIfEventsLogPluginSupported();
                 gerritConnection.addListener(missedEventsPlaybackManager);
 
-                setEventFilter(config.getFilterIn());
-
                 gerritConnection.start();
             } else {
                 logger.warn("Already started!");
@@ -611,16 +609,6 @@ public class GerritServer implements Describable<GerritServer>, Action {
     public void restartConnection() {
         stopConnection();
         startConnection();
-    }
-
-    /**
-     * Set the event type filter.
-     * @param eventFilter the event type filter
-     */
-    public void setEventFilter(List<String> eventFilter) {
-        if (gerritConnection != null) {
-            gerritConnection.setEventFilter(eventFilter);
-        }
     }
 
     /**
