+++ src/main/java/com/sonyericsson/hudson/plugins/gerrit/trigger/config/IGerritHudsonTriggerConfig.java
@@ -325,18 +325,4 @@ public interface IGerritHudsonTriggerConfig extends GerritConnectionConfig2 {
       * @return the instance of {@link Secret}.
       */
      Secret getGerritAuthKeyFileSecretPassword();
-
-     /**
-      * Get the number of events that are supported.
-      *
-      * @return the size of gerrit event type enum.
-      */
-     int getEventTypesSize();
-
-     /**
-      * Get the list of events that are filtered in.
-      *
-      * @return the list of events that are filtered in, if all events are included then it will return null.
-      */
-     List<String> getFilterIn();
 }
